- Made audio waveforms realise the new shareable behaviour. - Added a couple of new waveforms. - Removed waveform modifiers (shaper, rectifier, mutator) in favour of implementing faster direct realisations. - Moved antialiasing properties into a trait.
