+++ src/audio/signal/waveform/AliasedSaw.php
@@ -20,7 +20,7 @@ declare(strict_types=1);
 
 namespace ABadCafe\PDE\Audio\Signal\Waveform;
 use ABadCafe\PDE\Audio\Signal;
-
+use ABadCafe\PDE\Util;
 use function \ceil;
 
 /**
@@ -32,6 +32,8 @@ use function \ceil;
  * @see https://github.com/0xABADCAFE/random-proto-synth
  */
 class AliasedSaw implements Signal\IWaveform {
+
+    use Util\TAlwaysShareable;
 
     /**
      * Waveform period (interval after which it repeats).
