+++ src/audio/signal/waveform/Factory.php
@@ -39,8 +39,6 @@ class Factory implements Audio\IFactory {
         'Square'    => 'createSimple',
         'Noise'     => 'createSimple',
         'Pulse'     => 'createPulse',
-        'Rectifier' => 'createRectifier',
-        'Mutator'   => 'createMutator'
     ];
 
 
@@ -92,58 +90,5 @@ class Factory implements Audio\IFactory {
         return $bAliased ? new AliasedPulse() : new Pulse();
     }
 
-    /**
-     * Return a Rectifier based waveform.
-     *
-     * @param  \stdClass $oDefinition
-     * @param  string $sType
-     * @return Audio\Signal\IWaveform
-     */
-    private function createRectifier(\stdClass $oDefinition, $sType): Audio\Signal\IWaveform {
-        if (
-        empty($oDefinition->{self::STANDARD_KEY}) ||
-            !$oDefinition->{self::STANDARD_KEY} instanceof \stdClass
-        ) {
-            throw new \RuntimeException('Rectifier requires a waveform');
-        }
 
-        if (!empty($oDefinition->iPreset)) {
-            $iPreset = (int)$oDefinition->iPreset;
-            return Rectifier::createStandard($this->createFrom($oDefinition->{self::STANDARD_KEY}), $iPreset);
-        }
-
-        $fMinLevel = (float)($oDefinition->fMinLevel ?? -1.0);
-        $fMaxLevel = (float)($oDefinition->fMaxLevel ?? 1.0);
-        $fScale    = (float)($oDefinition->fScale    ?? 1.0);
-        $fBias     = (float)($oDefinition->fBias     ?? 0.0);
-        $bFold     = (bool)($oDefinition->bFold      ?? false);
-
-        return new Rectifier(
-            $this->createFrom($oDefinition->{self::STANDARD_KEY}),
-            $fMinLevel,
-            $fMaxLevel,
-            $bFold,
-            $fScale,
-            $fBias
-        );
-    }
-
-    /**
-     * Return a Mutator based waveform.
-     *
-     * @param  \stdClass $oDefinition
-     * @param  string $sType
-     * @return Audio\Signal\IWaveform
-     */
-    private function createMutator(\stdClass $oDefinition, $sType): Audio\Signal\IWaveform {
-        if (
-            empty($oDefinition->{self::STANDARD_KEY}) ||
-            !$oDefinition->{self::STANDARD_KEY} instanceof \stdClass
-        ) {
-            throw new \RuntimeException('Mutator requires a waveform');
-        }
-        return new QuadrantMutator(
-            $this->createFrom($oDefinition->{self::STANDARD_KEY})
-        );
-    }
 }
