+++ src/audio/signal/waveform/Triangle.php
@@ -20,7 +20,7 @@ declare(strict_types=1);
 
 namespace ABadCafe\PDE\Audio\Signal\Waveform;
 use ABadCafe\PDE\Audio\Signal;
-
+use ABadCafe\PDE\Util;
 use function \floor;
 
 /**
@@ -31,6 +31,8 @@ use function \floor;
  * @see https://github.com/0xABADCAFE/random-proto-synth
  */
 class Triangle implements Signal\IWaveform {
+
+    use Util\TAlwaysShareable;
 
     /**
      * Waveform period (interval after which it repeats).
