+++ src/audio/signal/waveform/sine/Pokey.php
@@ -19,46 +19,24 @@
 declare(strict_types=1);
 
 namespace ABadCafe\PDE\Audio\Signal\Waveform;
-use ABadCafe\PDE\Audio\Signal;
-use ABadCafe\PDE\Util;
-use function \floor;
 
 /**
- * Triangle
+ * Pokey
  *
- * Triangle implementation of IWaveform
+ * Fake Pokey implementation of IWaveform
  *
  * @see https://github.com/0xABADCAFE/random-proto-synth
  */
-class TriangleHalfRect implements Signal\IWaveform {
+class Pokey extends SineXForm {
 
-    use Util\TAlwaysShareable;
-
-    /**
-     * Waveform period (interval after which it repeats).
-     */
     const PERIOD = 2.0;
 
-    /**
-     * @inheritDoc
-     */
-    public function getPeriod(): float {
-        return self::PERIOD;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function map(Signal\Packet $oInput): Signal\Packet {
-        $oOutput = clone $oInput;
-        $fHalf   = 0.5;
-        foreach ($oInput as $i => $fTime) {
-            $fTime   -= $fHalf;
-            $fFloor  = floor($fTime);
-            $fScale  = (int)$fFloor & 1 ? 4.0 : -4.0;
-            $fSample = $fScale * ($fTime - $fFloor - $fHalf);
-            $oOutput[$i] = ($fSample > 0 ? $fSample : 0.0) - 1.0;
-        }
-        return $oOutput;
-    }
+    const TRANSFORM = [
+        // Quadrant phase shift, Bias Adjust, Scale.
+        // This default configuration rearranges a sine wave into something resembling a soft saw wave.
+        [ 2.0,  1.0, 0.25],
+        [ 1.0,  -1.0, -0.25],
+        [ 0.0,  1.0, 0.25],
+        [ -1.0, -1.0, -0.25]
+    ];
 }
