+++ src/audio/signal/waveform/sine/Sine.php
@@ -19,23 +19,41 @@
 declare(strict_types=1);
 
 namespace ABadCafe\PDE\Audio\Signal\Waveform;
-use ABadCafe\PDE\Audio;
 use ABadCafe\PDE\Audio\Signal;
-
+use ABadCafe\PDE\Util;
 use function \sin;
 
 /**
- * SineCut
+ * Sine
+ *
+ * Vanilla sine wave implementation of IWaveform
  *
  * @see https://github.com/0xABADCAFE/random-proto-synth
  */
-class SineCut extends SinePinch {
+class Sine implements Signal\IWaveform {
 
-    const TRANSFORM = [
-        // Quadrant phase shift, Bias Adjust, Scale.
-        [ 0.0, -1.0, 1.0],
-        [ 0.0, -1.0, 1.0],
-        [ 0.0,  1.0, 1.0],
-        [ 0.0,  1.0, 1.0]
-    ];
+    use Util\TAlwaysShareable;
+
+    /**
+     * Waveform period (interval after which it repeats).
+     */
+    const PERIOD = 2.0 * M_PI;
+
+    /**
+     * @inheritDoc
+     */
+    public function getPeriod(): float {
+        return self::PERIOD;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function map(Signal\Packet $oInput): Signal\Packet {
+        $oOutput = clone $oInput;
+        foreach ($oInput as $i => $fTime) {
+            $oOutput[$i] = sin($fTime); // @phpstan-ignore-line - false positive
+        }
+        return $oOutput;
+    }
 }
