+++ src/audio/signal/waveform/sine/SineCut.php
@@ -19,40 +19,19 @@
 declare(strict_types=1);
 
 namespace ABadCafe\PDE\Audio\Signal\Waveform;
-use ABadCafe\PDE\Audio\Signal;
-
-use function \sin;
 
 /**
- * SineFullRect
- *
- * Sinrwave implementation of IWaveform
+ * SineCut
  *
  * @see https://github.com/0xABADCAFE/random-proto-synth
  */
-class SineFullRect implements Signal\IWaveform {
+class SineCut extends SineXForm {
 
-    /**
-     * Waveform period (interval after which it repeats).
-     */
-    const PERIOD = M_PI;
-
-    /**
-     * @inheritDoc
-     */
-    public function getPeriod(): float {
-        return self::PERIOD;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function map(Signal\Packet $oInput): Signal\Packet {
-        $oOutput = clone $oInput;
-        foreach ($oInput as $i => $fTime) {
-            $fSin = sin($fTime); // @phpstan-ignore-line - false positive
-            $oOutput[$i] = 2.0*($fSin > 0.0 ? $fSin : -$fSin) - 1.0;
-        }
-        return $oOutput;
-    }
+    const TRANSFORM = [
+        // Quadrant phase shift, Bias Adjust, Scale.
+        [ 0.0, -1.0, 1.0],
+        [ 0.0, -1.0, 1.0],
+        [ 0.0,  1.0, 1.0],
+        [ 0.0,  1.0, 1.0]
+    ];
 }
