+++ src/audio/signal/waveform/sine/SineHalfRect.php
@@ -19,26 +19,42 @@
 declare(strict_types=1);
 
 namespace ABadCafe\PDE\Audio\Signal\Waveform;
-use ABadCafe\PDE\Audio;
 use ABadCafe\PDE\Audio\Signal;
-
+use ABadCafe\PDE\Util;
 use function \sin;
 
 /**
- * SineSaw
+ * SineHalfRect
  *
  * Sinrwave implementation of IWaveform
  *
  * @see https://github.com/0xABADCAFE/random-proto-synth
  */
-class SineSaw extends SinePinch {
+class SineHalfRect implements Signal\IWaveform {
 
-    const TRANSFORM = [
-        // Quadrant phase shift, Bias Adjust, Scale.
-        // This default configuration rearranges a sine wave into something resembling a triangle.
-        [ 3.0,  1.0, 1.0],
-        [ 0.0,  0.0, 1.0],
-        [ 0.0,  0.0, 1.0],
-        [-3.0,  -1.0, 1.0]
-    ];
+    use Util\TAlwaysShareable;
+
+    /**
+     * Waveform period (interval after which it repeats).
+     */
+    const PERIOD = 2.0 * M_PI;
+
+    /**
+     * @inheritDoc
+     */
+    public function getPeriod(): float {
+        return self::PERIOD;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function map(Signal\Packet $oInput): Signal\Packet {
+        $oOutput = clone $oInput;
+        foreach ($oInput as $i => $fTime) {
+            $fSin = sin($fTime); // @phpstan-ignore-line - false positive
+            $oOutput[$i] = ($fSin > 0 ? 2.0 * $fSin : 0.0) - 1.0;
+        }
+        return $oOutput;
+    }
 }
